requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
async-timeout==4.0.2; python_version >= '3.6'
asyncio==3.4.3
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
beautifulsoup4==4.11.1; python_version >= '3.6'
bs4==0.0.1
certifi==2022.5.18.1; python_version >= '3.6'
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3.5'
colorlog==6.6.0
discord-together==1.2.5
dnspython==2.2.1
ffmpeg-python==0.2.0
frozenlist==1.3.0; python_version >= '3.7'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3.5'
lxml==4.9.0
multidict==6.0.2; python_version >= '3.7'
novus==0.1.4
psutil==5.9.1
pycparser==2.21
pymongo[srv]==4.1.1
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.0
rule34==1.8.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
yarl==1.7.2; python_version >= '3.6'
